refactor(api-keys): extract RotateApiKeyDialog from action menu
ApiKeyActionMenu had grown to four dialogs in one file. Pull rotate
into its own component, mirroring how ExtendExpirationDialog already
lives separately. The dialog now owns its own constraints fetch,
last-used formatter, and rotation API call instead of leaking that
state up into the menu.
