+++ frontend/apps/web/src/lib/features/api-keys/ApiKeyActionMenu.svelte
@@ -25,10 +25,9 @@
   import * as Field from "$lib/components/ui/field/index.js";
   import * as Alert from "$lib/components/ui/alert/index.js";
   import { Button } from "$lib/components/ui/button/index.js";
-  import { Checkbox } from "$lib/components/ui/checkbox/index.js";
   import { Input } from "$lib/components/ui/input/index.js";
-  import ExpirationPicker from "$lib/features/api-keys/ExpirationPicker.svelte";
   import ExtendExpirationDialog from "$lib/features/api-keys/ExtendExpirationDialog.svelte";
+  import RotateApiKeyDialog from "$lib/features/api-keys/RotateApiKeyDialog.svelte";
 
   const intric = getIntric();
 
@@ -64,11 +63,6 @@
   let errorMessage = $state<string | null>(null);
   let reasonText = $state("");
   let followLoading = $state(false);
-  let rotationGraceHours = $state(24);
-  let rotateAlsoExtend = $state(false);
-  let rotateNewExpiresAt = $state<string | null>(null);
-  let rotateMaxDays = $state<number | null>(null);
-  let rotateRequireExpiration = $state(false);
 
   const isActive = $derived(apiKey.state === "active");
   const isSuspended = $derived(apiKey.state === "suspended");
@@ -85,56 +79,6 @@
       apiKey.ownership === "service" ||
       apiKey.owner_user_id === currentUserId
   );
-
-  function formatLastUsed(lastUsedAt: string | null | undefined): string | null {
-    if (!lastUsedAt) return null;
-    const last = new Date(lastUsedAt);
-    const now = new Date();
-    const diffMs = now.getTime() - last.getTime();
-    const diffMin = Math.floor(diffMs / 60_000);
-    if (diffMin < 1) return m.api_keys_rotate_last_used_just_now();
-    if (diffMin < 60) return m.api_keys_rotate_last_used_minutes({ minutes: diffMin });
-    const diffHours = Math.floor(diffMin / 60);
-    if (diffHours < 24) return m.api_keys_rotate_last_used_hours({ hours: diffHours });
-    const diffDays = Math.floor(diffHours / 24);
-    return m.api_keys_rotate_last_used_days({ days: diffDays });
-  }
-
-  async function openRotateDialog() {
-    rotateAlsoExtend = false;
-    rotateNewExpiresAt = apiKey.expires_at ?? null;
-    try {
-      const constraints = await intric.apiKeys.getCreationConstraints();
-      rotationGraceHours = constraints.rotation_grace_hours ?? 24;
-      rotateMaxDays = constraints.max_expiration_days ?? null;
-      rotateRequireExpiration = constraints.require_expiration ?? false;
-    } catch {
-      // Fall back to defaults
-    }
-    showRotateDialog = true;
-  }
-
-  async function rotateKey() {
-    actionPending = true;
-    try {
-      const params = rotateAlsoExtend
-        ? { id: apiKey.id, update_expiration: true, expires_at: rotateNewExpiresAt }
-        : { id: apiKey.id };
-      const response = isAdmin
-        ? await intric.apiKeys.admin.rotate(params)
-        : await intric.apiKeys.rotate(params);
-      if (!response?.secret) {
-        throw new Error("rotate_missing_secret");
-      }
-      showRotateDialog = false;
-      onSecret(response);
-    } catch (error) {
-      console.error(error);
-      toast.error(getErrorMessage(error));
-    } finally {
-      actionPending = false;
-    }
-  }
 
   async function revokeKey() {
     errorMessage = null;
@@ -247,7 +191,7 @@
       {/if}
 
       {#if canRotate}
-        <DropdownMenu.Item onclick={openRotateDialog} disabled={!canManage}>
+        <DropdownMenu.Item onclick={() => (showRotateDialog = true)} disabled={!canManage}>
           <RotateCcw />
           {isAdmin ? m.api_keys_admin_action_rotate() : m.api_keys_action_rotate()}
         </DropdownMenu.Item>
@@ -394,88 +338,6 @@
   </Dialog.Content>
 </Dialog.Root>
 
-<!-- Rotate confirmation dialog -->
-<Dialog.Root bind:open={showRotateDialog}>
-  <Dialog.Content class="sm:max-w-md">
-    <Dialog.Header>
-      <Dialog.Title>
-        {m.api_keys_rotate_confirm_title()}
-      </Dialog.Title>
-      <Dialog.Description>
-        {m.api_keys_rotate_confirm_description()}
-      </Dialog.Description>
-    </Dialog.Header>
-
-    <div class="space-y-3">
-      <div class="bg-subtle border-default rounded-lg border p-3">
-        <p class="text-default text-sm font-medium">{apiKey.name}</p>
-        <p class="text-muted mt-0.5 font-mono text-xs">
-          {apiKey.key_prefix}...{apiKey.key_suffix}
-        </p>
-      </div>
-
-      <div class="bg-subtle border-default space-y-1.5 rounded-lg border p-3">
-        <div class="flex items-center justify-between">
-          <span class="text-muted text-xs">{m.api_keys_rotate_grace_period_label()}</span>
-          <span class="text-default text-sm font-medium">
-            {m.api_keys_rotate_grace_period_value({ hours: rotationGraceHours })}
-          </span>
-        </div>
-        {#if apiKey.last_used_at}
-          {@const lastUsedText = formatLastUsed(apiKey.last_used_at)}
-          {#if lastUsedText}
-            <div class="flex items-center justify-between">
-              <span class="text-muted text-xs">{m.api_keys_rotate_last_used_label()}</span>
-              <span class="text-default text-sm">{lastUsedText}</span>
-            </div>
-          {/if}
-        {/if}
-        {#if apiKey.expires_at}
-          <div class="flex items-center justify-between">
-            <span class="text-muted text-xs">{m.api_keys_rotate_expires_label()}</span>
-            <span class="text-default text-sm">
-              {new Date(apiKey.expires_at).toLocaleDateString()}
-            </span>
-          </div>
-        {/if}
-      </div>
-
-      <Alert.Root>
-        <AlertCircle />
-        <Alert.Description class="text-xs">
-          {m.api_keys_rotate_grace_info({ hours: rotationGraceHours })}
-        </Alert.Description>
-      </Alert.Root>
-
-      <label class="flex items-center gap-2 text-sm">
-        <Checkbox
-          checked={rotateAlsoExtend}
-          onCheckedChange={(v) => (rotateAlsoExtend = v === true)}
-        />
-        <span>{m.api_keys_rotate_also_extend_label()}</span>
-      </label>
-
-      {#if rotateAlsoExtend}
-        <ExpirationPicker
-          bind:value={rotateNewExpiresAt}
-          maxDays={rotateMaxDays}
-          requireExpiration={rotateRequireExpiration}
-          disabled={actionPending}
-        />
-      {/if}
-    </div>
-
-    <Dialog.Footer>
-      <Dialog.Close>
-        {#snippet child({ props })}
-          <Button variant="outline" {...props}>{m.cancel()}</Button>
-        {/snippet}
-      </Dialog.Close>
-      <Button onclick={rotateKey} disabled={actionPending}>
-        {isAdmin ? m.api_keys_admin_action_rotate() : m.api_keys_action_rotate()}
-      </Button>
-    </Dialog.Footer>
-  </Dialog.Content>
-</Dialog.Root>
+<RotateApiKeyDialog {apiKey} {mode} bind:open={showRotateDialog} {onSecret} />
 
 <ExtendExpirationDialog {apiKey} {mode} bind:open={showExtendDialog} {onChanged} />
